+++ previous-versions/accounting/src/sdk/models/shared/transfer.ts
@@ -3,32 +3,13 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { DataType } from "./datatype";
+import { ContactRef } from "./contactref";
 import { InvoiceTo } from "./invoiceto";
 import { Metadata } from "./metadata";
 import { SupplementalData } from "./supplementaldata";
 import { TrackingCategoryRef } from "./trackingcategoryref";
 import { TransferAccount } from "./transferaccount";
 import { Expose, Type } from "class-transformer";
-
-/**
- * The customer or supplier for the transfer, if available.
- */
-export class TransferContactRef extends SpeakeasyBase {
-    /**
-     * Available Data types
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "dataType" })
-    dataType?: DataType;
-
-    /**
-     * Unique identifier for a customer or supplier.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "id" })
-    id: string;
-}
 
 /**
  * > View the coverage for transfers in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=transfers" target="_blank">Data coverage explorer</a>.
@@ -39,12 +20,12 @@ export class TransferContactRef extends SpeakeasyBase {
  */
 export class Transfer extends SpeakeasyBase {
     /**
-     * The customer or supplier for the transfer, if available.
+     * A customer or supplier associated with the direct cost.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "contactRef" })
-    @Type(() => TransferContactRef)
-    contactRef?: TransferContactRef;
+    @Type(() => ContactRef)
+    contactRef?: ContactRef;
 
     /**
      * In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example: